Features for global alloc and panic handler
Supersedes `init!` macro.
