+++ qf-polkavm-sdk/Cargo.toml
@@ -4,3 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+
+[features]
+global-allocator = []
+panic-handler = []
